chore: move genesis block builder to sc-chain-spec
