+++ client/block-builder/Cargo.toml
@@ -17,14 +17,13 @@ codec = { package = "parity-scale-codec", version = "3.2.2", features = [
 	"derive",
 ] }
 sc-client-api = { version = "4.0.0-dev", path = "../api" }
-sc-executor = { version = "0.10.0-dev", path = "../executor" }
 sp-api = { version = "4.0.0-dev", path = "../../primitives/api" }
 sp-block-builder = { version = "4.0.0-dev", path = "../../primitives/block-builder" }
 sp-blockchain = { version = "4.0.0-dev", path = "../../primitives/blockchain" }
 sp-core = { version = "7.0.0", path = "../../primitives/core" }
 sp-inherents = { version = "4.0.0-dev", path = "../../primitives/inherents" }
 sp-runtime = { version = "7.0.0", path = "../../primitives/runtime" }
-sp-state-machine = { version = "0.13.0", path = "../../primitives/state-machine" }
 
 [dev-dependencies]
+sp-state-machine = { version = "0.13.0", path = "../../primitives/state-machine" }
 substrate-test-runtime-client = { path = "../../test-utils/runtime/client" }
