+++ client/block-builder/src/lib.rs
@@ -26,11 +26,6 @@
 
 #![warn(missing_docs)]
 
-mod genesis;
-
-pub use genesis::{construct_genesis_block, BuildGenesisBlock, GenesisBlockBuilder};
-use std::collections::hash_map::DefaultHasher;
-
 use codec::Encode;
 
 use sp_api::{
@@ -46,41 +41,7 @@ use sp_runtime::{
 };
 
 use sc_client_api::backend;
-use sc_executor::RuntimeVersionOf;
 pub use sp_block_builder::BlockBuilder as BlockBuilderApi;
-use sp_core::storage::{well_known_keys, StateVersion, Storage};
-
-/// Return the state version given the genesis storage and executor.
-pub fn resolve_state_version_from_wasm<E>(
-	storage: &Storage,
-	executor: &E,
-) -> sp_blockchain::Result<StateVersion>
-where
-	E: RuntimeVersionOf,
-{
-	if let Some(wasm) = storage.top.get(well_known_keys::CODE) {
-		let mut ext = sp_state_machine::BasicExternalities::new_empty(); // just to read runtime version.
-
-		let code_fetcher = sp_core::traits::WrappedRuntimeCode(wasm.as_slice().into());
-		let runtime_code = sp_core::traits::RuntimeCode {
-			code_fetcher: &code_fetcher,
-			heap_pages: None,
-			hash: {
-				use std::hash::{Hash, Hasher};
-				let mut state = DefaultHasher::new();
-				wasm.hash(&mut state);
-				state.finish().to_le_bytes().to_vec()
-			},
-		};
-		let runtime_version = RuntimeVersionOf::runtime_version(executor, &mut ext, &runtime_code)
-			.map_err(|e| sp_blockchain::Error::VersionInvalid(e.to_string()))?;
-		Ok(runtime_version.state_version())
-	} else {
-		Err(sp_blockchain::Error::VersionInvalid(
-			"Runtime missing from initial storage, could not read state version.".to_string(),
-		))
-	}
-}
 
 /// Used as parameter to [`BlockBuilderProvider`] to express if proof recording should be enabled.
 ///
