+++ client/chain-spec/Cargo.toml
@@ -16,8 +16,12 @@ targets = ["x86_64-unknown-linux-gnu"]
 memmap2 = "0.5.0"
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.85"
+sc-client-api = { version = "4.0.0-dev", path = "../api" }
 sc-chain-spec-derive = { version = "4.0.0-dev", path = "./derive" }
+sc-executor = { version = "0.10.0-dev", path = "../executor" }
 sc-network-common = { version = "0.10.0-dev", path = "../network/common" }
 sc-telemetry = { version = "4.0.0-dev", path = "../telemetry" }
+sp-blockchain = { version = "4.0.0-dev", path = "../../primitives/blockchain" }
 sp-core = { version = "7.0.0", path = "../../primitives/core" }
 sp-runtime = { version = "7.0.0", path = "../../primitives/runtime" }
+sp-state-machine = { version = "0.13.0", path = "../../primitives/state-machine" }
