+++ client/chain-spec/src/genesis.rs
@@ -18,21 +18,56 @@
 
 //! Tool for creating the genesis block.
 
-use std::{marker::PhantomData, sync::Arc};
+use std::{collections::hash_map::DefaultHasher, marker::PhantomData, sync::Arc};
 
 use sc_client_api::{backend::Backend, BlockImportOperation};
 use sc_executor::RuntimeVersionOf;
-use sp_core::storage::Storage;
+use sp_core::storage::{well_known_keys, StateVersion, Storage};
 use sp_runtime::{
 	traits::{Block as BlockT, Hash as HashT, Header as HeaderT, Zero},
 	BuildStorage,
 };
 
+/// Return the state version given the genesis storage and executor.
+pub fn resolve_state_version_from_wasm<E>(
+	storage: &Storage,
+	executor: &E,
+) -> sp_blockchain::Result<StateVersion>
+where
+	E: RuntimeVersionOf,
+{
+	if let Some(wasm) = storage.top.get(well_known_keys::CODE) {
+		let mut ext = sp_state_machine::BasicExternalities::new_empty(); // just to read runtime version.
+
+		let code_fetcher = sp_core::traits::WrappedRuntimeCode(wasm.as_slice().into());
+		let runtime_code = sp_core::traits::RuntimeCode {
+			code_fetcher: &code_fetcher,
+			heap_pages: None,
+			hash: {
+				use std::hash::{Hash, Hasher};
+				let mut state = DefaultHasher::new();
+				wasm.hash(&mut state);
+				state.finish().to_le_bytes().to_vec()
+			},
+		};
+		let runtime_version = RuntimeVersionOf::runtime_version(executor, &mut ext, &runtime_code)
+			.map_err(|e| sp_blockchain::Error::VersionInvalid(e.to_string()))?;
+		Ok(runtime_version.state_version())
+	} else {
+		Err(sp_blockchain::Error::VersionInvalid(
+			"Runtime missing from initial storage, could not read state version.".to_string(),
+		))
+	}
+}
+
 /// Create a genesis block, given the initial storage.
-pub fn construct_genesis_block<Block: BlockT>(state_root: Block::Hash) -> Block {
+pub fn construct_genesis_block<Block: BlockT>(
+	state_root: Block::Hash,
+	state_version: StateVersion,
+) -> Block {
 	let extrinsics_root = <<<Block as BlockT>::Header as HeaderT>::Hashing as HashT>::trie_root(
 		Vec::new(),
-		sp_runtime::StateVersion::V0,
+		state_version,
 	);
 
 	Block::new(
@@ -94,12 +129,11 @@ impl<Block: BlockT, B: Backend<Block>, E: RuntimeVersionOf> BuildGenesisBlock<Bl
 	fn build_genesis_block(self) -> sp_blockchain::Result<(Block, Self::BlockImportOperation)> {
 		let Self { genesis_storage, commit_genesis_state, backend, executor, _phantom } = self;
 
-		let genesis_state_version =
-			crate::resolve_state_version_from_wasm(&genesis_storage, &executor)?;
+		let genesis_state_version = resolve_state_version_from_wasm(&genesis_storage, &executor)?;
 		let mut op = backend.begin_operation()?;
 		let state_root =
 			op.set_genesis_state(genesis_storage, commit_genesis_state, genesis_state_version)?;
-		let genesis_block = construct_genesis_block::<Block>(state_root);
+		let genesis_block = construct_genesis_block::<Block>(state_root, StateVersion::V0);
 
 		Ok((genesis_block, op))
 	}
