+++ client/chain-spec/src/lib.rs
@@ -177,10 +177,15 @@
 
 mod chain_spec;
 mod extension;
+mod genesis;
 
-pub use chain_spec::{ChainSpec as GenericChainSpec, NoExtension};
-pub use extension::{
-	get_extension, get_extension_mut, Extension, Fork, Forks, GetExtension, Group,
+pub use self::{
+	chain_spec::{ChainSpec as GenericChainSpec, NoExtension},
+	extension::{get_extension, get_extension_mut, Extension, Fork, Forks, GetExtension, Group},
+	genesis::{
+		construct_genesis_block, resolve_state_version_from_wasm, BuildGenesisBlock,
+		GenesisBlockBuilder,
+	},
 };
 pub use sc_chain_spec_derive::{ChainSpecExtension, ChainSpecGroup};
