+++ test-utils/runtime/client/Cargo.toml
@@ -15,6 +15,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 codec = { package = "parity-scale-codec", version = "3.2.2" }
 futures = "0.3.21"
 sc-block-builder = { version = "0.10.0-dev", path = "../../../client/block-builder" }
+sc-chain-spec = { version = "4.0.0-dev", path = "../../../client/chain-spec" }
 sc-client-api = { version = "4.0.0-dev", path = "../../../client/api" }
 sc-consensus = { version = "0.10.0-dev", path = "../../../client/consensus/common" }
 sp-api = { version = "4.0.0-dev", path = "../../../primitives/api" }
