+++ test-utils/runtime/client/src/lib.rs
@@ -37,6 +37,8 @@ use sp_core::{
 };
 use sp_runtime::traits::{Block as BlockT, Hash as HashT, Header as HeaderT};
 use substrate_test_runtime::genesismap::{additional_storage_with_genesis, GenesisConfig};
+use sc_chain_spec::construct_genesis_block;
+use sp_api::StateVersion;
 
 /// A prelude to import in tests.
 pub mod prelude {
@@ -122,7 +124,7 @@ impl GenesisParameters {
 	}
 }
 
-impl substrate_test_client::GenesisInit for GenesisParameters {
+impl GenesisInit for GenesisParameters {
 	fn genesis_storage(&self) -> Storage {
 		use codec::Encode;
 
@@ -148,7 +150,7 @@ impl substrate_test_client::GenesisInit for GenesisParameters {
 				storage.top.clone().into_iter().chain(child_roots).collect(),
 				sp_runtime::StateVersion::V1,
 			);
-		let block: runtime::Block = client::genesis::construct_genesis_block(state_root);
+		let block: runtime::Block = construct_genesis_block(state_root, StateVersion::V0);
 		storage.top.extend(additional_storage_with_genesis(&block));
 
 		storage
@@ -260,7 +262,7 @@ impl<B> TestClientBuilderExt<B>
 		client::LocalCallExecutor<
 			substrate_test_runtime::Block,
 			B,
-			sc_executor::NativeElseWasmExecutor<LocalExecutorDispatch>,
+			NativeElseWasmExecutor<LocalExecutorDispatch>,
 		>,
 		B,
 	> where
@@ -288,8 +290,8 @@ pub fn new() -> Client<Backend> {
 }
 
 /// Create a new native executor.
-pub fn new_native_executor() -> sc_executor::NativeElseWasmExecutor<LocalExecutorDispatch> {
-	sc_executor::NativeElseWasmExecutor::new(
+pub fn new_native_executor() -> NativeElseWasmExecutor<LocalExecutorDispatch> {
+	NativeElseWasmExecutor::new(
 		sc_executor::WasmExecutionMethod::Interpreted,
 		None,
 		8,
